всё, кроме раздельной сортировки

+++ _compl.py
@@ -5,12 +5,29 @@ def complgen(perecod):
     import csv
     import sys
     import os
+
+    if not os.path.exists(os.path.abspath('csv_compl.csv')):
+        print 'FATAL ERROR!!! \n' 
+        ofile =open('reports.tex', 'w')
+        ofile.write('&&&'+'file csv compl ERROR!!!'+'&&&'+'\\'+'\\''\n')
+        ofile.close()
+        sys.exit()
     
     ifile  = open('csv_compl.csv', "rb")
     readerd = csv.DictReader(ifile, delimiter=";", quoting=csv.QUOTE_NONE)
     ofile =open('reports_compl.tex', 'w')
     row_num = 0
     for row in readerd:
+        if row_num==0:
+            headerd=row
+            if not (('Kol' in headerd) and ('Oboz' in headerd) and ('Name' in headerd)):
+                print 'FATAL ERROR!!! \n'
+                ofile.close()                        
+                ofile =open('reports.tex', 'w')
+                ofile.write('&&&'+'file csv compl ERROR!!!'+'&&&'+'\\'+'\\''\n')
+                ifile.close()
+                ofile.close()
+                sys.exit()
         row_num += 1
     if row_num > 0:
         header=row
