+++ _det.py
@@ -6,12 +6,30 @@ def detgen(num,perecod):
     import sys
     import os
     
+    if not os.path.exists(os.path.abspath('csv_det.csv')):
+        print 'FATAL ERROR!!! \n' 
+        ofile =open('reports.tex', 'w')
+        ofile.write('&&&'+'file csv det ERROR!!!'+'&&&'+'\\'+'\\''\n')
+        ofile.close()
+        sys.exit()
+        
     ifile  = open('csv_det.csv', "rb")
     readerd = csv.DictReader(ifile, delimiter=";", quoting=csv.QUOTE_NONE)
     ofile =open('reports_det.tex', 'w')
     row_num = 0
     raz = 0
     for row in readerd:
+        if row_num==0:
+            headerd=row
+            if not (('Form' in headerd) and ('Oboz' in headerd) and ('Name' in headerd)
+                     and ('Poz' in headerd) and ('Kol' in headerd)):
+                print 'FATAL ERROR!!! \n'
+                ofile.close()                        
+                ofile =open('reports.tex', 'w')
+                ofile.write('&&&'+'file csv det ERROR!!!'+'&&&'+'\\'+'\\''\n')
+                ifile.close()
+                ofile.close()
+                sys.exit()
         row_num += 1
     if row_num > 0:
         header=row
